restrict edit and delete image functionality to owner only

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -35,6 +35,17 @@ public class ImageController {
         return "images";
     }
 
+    //method to check if current logged in user is creator of the image
+
+    private Boolean checkImagePermission (HttpSession session, Integer imageId){
+        User currentUser = (User) session.getAttribute("loggeduser");
+        if( currentUser.getId()  ==  imageService.getImage(imageId).getUser().getId()){
+            return true;
+        }
+        return false;
+    }
+
+
     //This method is called when the details of the specific image with corresponding title are to be displayed
     //The logic is to get the image from the databse with corresponding title. After getting the image from the database the details are shown
     //First receive the dynamic parameter in the incoming request URL in a string variable 'title' and also the Model type object
@@ -95,10 +106,22 @@ public class ImageController {
     public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
 
+        if(checkImagePermission(session, imageId) ) {
+
             String tags = convertTagsToString(image.getTags());
             model.addAttribute("image", image);
             model.addAttribute("tags", tags);
             return "images/edit";
+        }
+        else{
+            String imageTitle = imageService.getImage(imageId).getTitle();
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,13 +163,24 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
 
         Image image = imageService.getImage(imageId);
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
-    }
 
+        if(checkImagePermission(session, imageId) )
+        {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else{
+            String imageTitle = imageService.getImage(imageId).getTitle();
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
+    }
 
     //This method converts the image to Base64 format
     private String convertUploadedFileToBase64(MultipartFile file) throws IOException {
